Fix TCB struct alignment.
	* sysdeps/powerpc/nptl/tls.h: Corrected hwcap field to 64 bit.
	* sysdeps/powerpc/sys/platform/ppc.h: Corrected TCB offsets to
	reflect changes.

+++ sysdeps/powerpc/nptl/tls.h
@@ -74,7 +74,7 @@ typedef struct
 #endif
   /* Reservation for HWCAP data. To be accessed by GCC in
      __builtin_cpu_supports(), so it is a part of public ABI.  */
-  unsigned int hwcap;
+  uint64_t hwcap;
   /* Indicate if HTM capable (ISA 2.07).  */
   int tm_capable;
   /* Reservation for AT_PLATFORM data - powerpc64.  */
